requirements.txt
# Python dependencies for data collection

# Core
python-dotenv==1.0.0
requests==2.31.0

# Google APIs
google-api-python-client==2.108.0
google-generativeai==0.3.2

# Data processing
pandas==2.1.4
openpyxl==3.1.2  # For Excel export

# Optional - Apify (if needed later)
# apify-client==1.5.0
